refactor: 세션 클래스 이동
