+++ tomcat/src/main/java/com/techcourse/handler/LoginHandler.java
@@ -5,7 +5,7 @@ import java.util.Map;
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
 import org.apache.coyote.http11.Cookie;
-import org.apache.coyote.http11.Session;
+import org.apache.catalina.Session;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
